# specs/004-timesheet-roles/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: 報工系統 API - 角色權限管理
  description: |
    報工系統的 RESTful API，支援五種角色（管理層、PM、部門主管、執行人員、HR）的權限管理。
    所有 API（除了 /auth 端點）都需要JWT 認證。
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@example.com
servers:
  - url: http://localhost:8080/api
    description: 本地開發環境
  - url: https://api.timesheet.example.com/api
    description: 生產環境
...
security:
  - BearerAuth: []
paths:
...
  /employee/tasks/{id}/complete:
    patch:
      tags: [Employee]
      summary: 標記任務為完成
      description: 標記任務完成，剩餘時數將釋放回專案池
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 任務標記完成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'

  # ===== PM Endpoints =====
...
  /pm/tasks/{id}:
    put:
      tags: [PM]
      summary: 更新任務資訊
      description: 可更新任務名稱、描述、預估時數、指派人員
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: 無權編輯此任務
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [PM]
      summary: 刪除任務
      description: |
        僅能刪除沒有工時記錄的任務。
        如果任務已有工時記錄，系統會拒絕刪除並提示改用關閉功能。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 刪除成功
        '400':
          description: 任務有工時記錄，無法刪除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: TASK_HAS_TIMESHEETS
                message: 此任務已有工時記錄，請改用關閉功能
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: 張三
        email:
          type: string
          format: email
          example: zhangsan@example.com
        role:
          type: string
          enum: [EXECUTIVE, PM, MANAGER, EMPLOYEE, HR]
          example: PM
        department:
          $ref: '#/components/schemas/DepartmentDto'
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
...
    DepartmentDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: 研發部
        manager:
          $ref: '#/components/schemas/UserDto'

    # Project-related schemas
    ProjectDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: 客戶管理系統
        description:
          type: string
          example: 開發新的客戶管理系統
        totalHours:
          type: number
          format: double
          example: 1000.00
        usedHours:
          type: number
          format: double
          example: 350.50
        remainingHours:
          type: number
          format: double
          example: 649.50
        status:
          type: string
          enum: [ACTIVE, CLOSED]
          example: ACTIVE
        pm:
          $ref: '#/components/schemas/UserDto'
        createdBy:
          $ref: '#/components/schemas/UserDto'
        createdAt:
          type: string
          format: date-time
...
    TaskDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: 實作使用者登入功能
        description:
          type: string
          example: 開發前後端的使用者登入功能
        project:
          $ref: '#/components/schemas/ProjectDto'
        assignee:
          $ref: '#/components/schemas/UserDto'
        estimatedHours:
          type: number
          format: double
          example: 40.00
        usedHours:
          type: number
          format: double
          example: 25.50
        remainingHours:
          type: number
          format: double
          example: 14.50
        status:
          type: string
          enum: [IN_PROGRESS, COMPLETED, CLOSED, PENDING_REASSIGNMENT]
          example: IN_PROGRESS
        completedAt:
          type: string
          format: date-time
          nullable: true
        actualHours:
          type: number
          format: double
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    UpdateTaskRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 200
        description:
          type: string
        estimatedHours:
          type: number
          format: double
          minimum: 0
        assigneeId:
          type: integer
          format: int64

    # Timesheet-related schemas
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: 驗證失敗
        details:
          type: array
          items:
            type: string
          example: ["工時日期必須在過去三個工作天內"]
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: 請求格式錯誤或驗證失敗
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...